Working example

+++ Xcode2Rust Shared/Bridge.h
@@ -17,10 +17,10 @@
 
 void rust_draw(uint8_t *pixels, uint32_t width, uint32_t height);
 
-bool rust_mouse_down(uint32_t x, uint32_t y);
+bool rust_mouse_down(float x, float y);
 
-bool rust_mouse_dragged(uint32_t x, uint32_t y);
+bool rust_mouse_dragged(float x, float y);
 
-bool rust_mouse_up(uint32_t x, uint32_t y);
+bool rust_mouse_up(float x, float y);
 
 #endif /* Header_h */
